Added the following elements
a) LoginServlet: Handles user attempt to login. Only checks for the username for now. If it succeeds, then I intend the servlet to display the "MainMenu.jsp" page for that user. Also, this servlet begins a new session, where the logged in student is set as the user. If the servlet cannot retrieve a student by the given username, then request is forwarded to the "NotAuthorized.jsp" page.

b) RegistrationServlet: Handles user attempt to register. Checks if request.getParameter("action") == "add". If this condition isn't satisfied, then the "BadRequest.jsp" page is displayed, If this condition is satisfied, then control is passed on to the "UserManagementServlet".

c)UserManagementServlet: This servlet is only slightly modified to include an empty body, empty argument constructor.